Use sanitized user login when logging in after registration
If user does not submit a username, their email will be used.  If user's
username has invalid characters in it, wordpress will sanitize it inside
`wp_insert_user`.  In either of these cases, the workflow is very
confusing, as the user entry is created in the database, but the user is
not logged in.  Therefore, if username is successfully saved in
database, that same username should be used to log in