Raise an error if user does not provide password when registering
Otherwise the user account is created but we cannot log the user in,
which confusing scenario in which the user tries to re-register.  If we
were to generate a passwor using `wp_generate_password`, we would not be
able to retrieve it via `$user->user_pass` in order to log in.
Therefore, it is better to deny user creation when no password is
provided